Test de conversion string vers type primitif

+++ CSVReaderUI/Fruit.cs
@@ -12,8 +12,8 @@ namespace CSVReaderUI
     {
         public string Nom { get; set; }
         public string Provenance { get; set; }
-        public double Prix { get; set; }
-        public int Quantite { get; set; }
+        public decimal Prix { get; set; }
+        public float Quantite { get; set; }
         public string Inutile { get; set; }
     }
 }
